+++ lib/changelog/gem_version.rb
@@ -1,3 +1,3 @@
 module Changelog
-  VERSION = '0.7.1'.freeze
+  VERSION = '0.7.2'.freeze
 end